# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactDetailResponse": {
        "description": "Single-artifact view; ``meta_json`` only when requested.",
        "properties": {
          "analysis_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "blob_deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Deleted At"
          },
          "byte_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Byte Size"
          },
          "content_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Sha256"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluation_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Run Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "meta_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Json"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "role_key": {
            "title": "Role Key",
            "type": "string"
          },
          "run_step_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Step Id"
          },
          "storage_uri": {
            "description": "Backend-specific object locator (e.g. local:\u2026)",
            "title": "Storage Uri",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role_key",
          "kind",
          "storage_uri",
          "created_at",
          "updated_at"
        ],
        "title": "ArtifactDetailResponse",
        "type": "object"
      },
      "ArtifactKind": {
        "description": "Broad artifact classification for storage and UI routing.",
        "enum": [
          "tabular",
          "document",
          "binary",
          "json",
          "other"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
...
      "ArtifactPreviewResponse": {
        "description": "Bounded UTF-8 text (or pretty-printed JSON) for browser preview.\n\nBinary or non-previewable types return **415** from the preview route instead.",
        "properties": {
          "format": {
            "description": "``json`` when payload parsed as JSON; else ``text``.",
            "enum": [
              "text",
              "json"
            ],
            "title": "Format",
            "type": "string"
          },
          "json_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "When ``format`` is ``json``, whether parsing succeeded.",
            "title": "Json Valid"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Artifact MIME type from metadata.",
            "title": "Mime Type"
          },
          "text": {
            "description": "UTF-8 preview string (may be truncated).",
            "title": "Text",
            "type": "string"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Declared size from metadata when known.",
            "title": "Total Bytes"
          },
          "truncated": {
            "description": "True when more bytes exist beyond the server preview cap.",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "format",
          "text",
          "truncated",
          "mime_type"
        ],
        "title": "ArtifactPreviewResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "EDGAR Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "description": "When true, include meta_json (may be large)",
            "in": "query",
            "name": "include_meta",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include meta_json (may be large)",
              "title": "Include Meta",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/v1/artifacts/{artifact_id}/content": {
      "get": {
        "description": "Stream artifact bytes from the object store (``local:`` or future backends).\n\nUses :func:`backend.storage.resolver.open_reader` \u2014 no raw filesystem paths in the response.",
        "operationId": "get_artifact_content_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "description": "``auto``: inline for text/json-like MIME types, else attachment.",
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "default": "auto",
              "description": "``auto``: inline for text/json-like MIME types, else attachment.",
              "enum": [
                "inline",
                "attachment",
                "auto"
              ],
              "title": "Disposition",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact Content",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/v1/artifacts/{artifact_id}/preview": {
      "get": {
        "description": "Return a bounded UTF-8 text (or pretty JSON) preview for markdown/CSV/JSON/text artifacts.\n\nZero-byte objects return ``text: \"\"`` with ``truncated: false``. See ``docs/artifact-delivery.md``\nfor preview eligibility and response fields.\n\nReturns **415** when the artifact is not treated as text-previewable (e.g. binary).",
        "operationId": "get_artifact_preview_v1_artifacts__artifact_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact Preview",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/v1/demos/{slug}/artifacts/{artifact_id}/content": {
      "get": {
        "description": "Stream the bytes of an artifact behind a published demo.",
        "operationId": "get_demo_artifact_content_v1_demos__slug__artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "description": "``auto``: inline for text/json-like MIME types, else attachment.",
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "default": "auto",
              "description": "``auto``: inline for text/json-like MIME types, else attachment.",
              "enum": [
                "inline",
                "attachment",
                "auto"
              ],
              "title": "Disposition",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Demo Artifact Content",
        "tags": [
          "demos"
        ]
      }
    },
    "/v1/demos/{slug}/artifacts/{artifact_id}/preview": {
      "get": {
        "description": "Bounded text/JSON preview of an artifact behind a published demo.",
        "operationId": "get_demo_artifact_preview_v1_demos__slug__artifacts__artifact_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Demo Artifact Preview",
        "tags": [
          "demos"
        ]
      }
    },
...
  }
}